refactor: 사용자 큐 요청 처리 및 사용자 추출 기능 관심사 분리
